refactor(consumption): extract bad-scan report sections (Refs #563)
bad_scan_report_sheet.dart was 611 LOC. Split into focused
sub-widgets and pure formatters so the parent file fits the
<300 LOC ceiling while preserving the public API
(BadScanReportSheet, ScanKind, ConsentPrompter etc.).

- bad_scan_diff_table.dart — BadScanDiffRow + BadScanDiffTable
- bad_scan_issue_created_surface.dart — post-submission surface
- bad_scan_form_view.dart — pre-submission title + diff + actions
- bad_scan_report_formatters.dart — pure helpers (title resolution,
  diff rows, share body, parsedFields/userCorrections maps)
- bad_scan_report_sheet.dart — keeps state + submit handler only,
  now 296 LOC

No behaviour change. The 13 existing widget tests in
bad_scan_report_sheet_test.dart still pass unmodified.
